Cambio en los controladores y rutas
Moví los controladores en una carpeta, creando archivos para cada controlador y cambie las rutas teniendo en cuenta la ubicación de los archivos actuales.
