+++ BACKEND/controllers/AuthController/Logout.js
@@ -1,0 +1,4 @@
+exports.logout = (req, res)=>{
+    res.json({message: "logout exitoso"});
+  };
+
