+++ BACKEND/routes/product.js
@@ -1,20 +1,16 @@
 const express = require('express');
 const router = express.Router();
-const productController = require('../controllers/productController'); // Asegúrate de que la ruta al controlador sea correcta
+const { createProduct } = require("../controllers/ProductController/CreateProduct");
+const { deleteProduct } = require("../controllers/ProductController/DeleteProduct");
+const { getProduct } = require("../controllers/ProductController/GetProduct");
+const { getProducts } = require("../controllers/ProductController/GetProducts");
+const { updateProduct } = require("../controllers/ProductController/UpdateProduct");
 
-// Ruta para obtener todos los productos
-router.get('/', productController.getProducts);
 
-// Ruta para obtener un producto específico por ID
-router.get('/:id', productController.getProduct);
-
-// Ruta para crear un nuevo producto
-router.post('/', productController.createProduct);
-
-// Ruta para actualizar un producto existente
-router.put('/:id', productController.updateProduct);
-
-// Ruta para eliminar un producto (marcarlo como eliminado)
-router.delete('/:id', productController.deleteProduct);
+router.post("/", createProduct);                
+router.get("/", getProducts);                   
+router.get("/:id", getProduct);                 
+router.put("/:id", updateProduct);              
+router.delete("/:id", deleteProduct);           
 
 module.exports = router;
